更新，每日释放，新增两个后台操作接口，生态用户列表－生态释放流水

+++ models/SuperForceTable.go
@@ -4,7 +4,7 @@ package models
 type SuperForceTable struct {
 	Id               int     `orm:"column(id);pk;auto" json:"id"`
 	Level            string  `orm:"column(level)" json:"level" json:"level"`
-	CoinNumberRule   int     `orm:"column(coin_number_rule)" json:"coin_number_rule"`
+	CoinNumberRule   float64 `orm:"column(coin_number_rule)" json:"coin_number_rule"`
 	BonusCalculation float64 `orm:"column(bonus_calculation)" json:"bonus_calculation"`
 }
